+++ libs/forms-vue-shadcn/src/lib/uikit/primitives/checkbox.ts
@@ -1,8 +1,8 @@
 import type { UIKitCheckboxProps } from '@letar/forms-core/uikit'
+import { cn } from '@letar/tailwind-utils'
 import { Check } from 'lucide-vue-next'
 import { CheckboxIndicator, CheckboxRoot } from 'reka-ui'
 import { h, type VNode } from 'vue'
-import { cn } from '@letar/tailwind-utils'
 import type { UINode } from '../ui-node'
 
 export function Checkbox(