+++ libs/forms-vue-shadcn/src/lib/uikit/primitives/combobox.ts
@@ -1,4 +1,6 @@
 import type { UIKitComboboxProps } from '@letar/forms-core/uikit'
+import { cn } from '@letar/tailwind-utils'
+import { NATIVE_INPUT_CLASS } from '@letar/tailwind-utils'
 import { Check } from 'lucide-vue-next'
 import {
   ComboboxAnchor,
@@ -12,9 +14,7 @@ import {
   ComboboxViewport,
 } from 'reka-ui'
 import { h, type VNode } from 'vue'
-import { cn } from '@letar/tailwind-utils'
 import type { UINode } from '../ui-node'
-import { NATIVE_INPUT_CLASS } from '@letar/tailwind-utils'
 
 export function Combobox(
   { value, inputValue, onInputChange, onValueChange, options, loading, placeholder, disabled, ...rest }: