+++ libs/forms-vue-shadcn/src/lib/uikit/primitives/input.ts
@@ -1,7 +1,7 @@
 import type { UIKitInputProps } from '@letar/forms-core/uikit'
-import { h, type VNode } from 'vue'
 import { cn } from '@letar/tailwind-utils'
 import { NATIVE_INPUT_CLASS } from '@letar/tailwind-utils'
+import { h, type VNode } from 'vue'
 
 export function Input({
   value,